==> code/04_nlp_model.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns


from nltk.stem import WordNetLemmatizer

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import train_test_split, GridSearchCV

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier, ExtraTreesClassifier
from sklearn.pipeline import Pipeline

from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay

df_sub = pd.read_csv('../data/df_all_subm.csv')
df_com = pd.read_csv('../data/df_all_comm.csv')
df_all = pd.read_csv('../data/df_all_text.csv')

X_sub = df_sub['title']
y_sub = df_sub['is_news']

# split training set and final test set
XS_training, XS_final, yS_training, yS_final = train_test_split(X_sub, y_sub, test_size = 0.1, stratify = y_sub)

XS_training.shape, yS_training.shape, XS_final.shape, yS_final.shape

# repeat for comments df
X_com = df_com['body']
y_com = df_com['is_news']

XC_training, XC_final, yC_training, yC_final = train_test_split(X_com, y_com, test_size = 0.1, stratify = y_com)

XC_training.shape, yC_training.shape, XC_final.shape, yC_final.shape

# split training sets into train-test for initial modeling
XS_train, XS_test, yS_train, yS_test = train_test_split(XS_training, yS_training, stratify = yS_training)

XC_train, XC_test, yC_train, yC_test = train_test_split(XC_training, yC_training, stratify = yC_training)

def my_lemma(item):
    lemma = WordNetLemmatizer()
    split_items = item.split(' ')
    lemmatized = []
    for i in split_items:
        lemmatized.append(lemma.lemmatize(i))
    return lemmatized

Xa_training, Xa_final, ya_training, ya_final = train_test_split(X_a, y_a, test_size = 0.1, stratify = y_a)

Xa_train, Xa_test, ya_train, ya_test = train_test_split(Xa_training, ya_training, stratify = ya_training)

---
## Random Forest CVEC/TVEC

rfc_cvec_pipe = Pipeline([
    ('cvec', CountVectorizer()),
    ('rfc', RandomForestClassifier())
])

rfc_cvec_params = {
    'cvec__stop_words' : ['english', None],
    'cvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)],
    'rfc__n_estimators': [100, 200, 300]
}

gs_rfc_cvec = GridSearchCV(rfc_cvec_pipe, param_grid = rfc_cvec_params)

gs_rfc_cvec.fit(XS_train, yS_train)

gs_rfc_cvec.best_params_

gs_rfc_cvec.score(XS_train, yS_train), gs_rfc_cvec.score(XS_test, yS_test), gs_rfc_cvec.score(XS_final, yS_final)

---

rfc_tvec_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('rfc', RandomForestClassifier())
])

rfc_tvec_params = {
    'tfidf__stop_words': ['english', None],
    'tfidf__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)],
    'rfc__n_estimators': range(100, 501, 50)
}

gs_rfc_tvec = GridSearchCV(rfc_tvec_pipe, param_grid = rfc_tvec_params)

gs_rfc_tvec.fit(XS_train, yS_train)

gs_rfc_tvec.best_params_

gs_rfc_tvec.score(XS_train, yS_train), gs_rfc_tvec.score(XS_test, yS_test), gs_rfc_tvec.score(XS_final, yS_final)

gs_rfc_tvec.score(XC_train, yC_train), gs_rfc_tvec.score(XC_test, yC_test), gs_rfc_tvec.score(XC_final, yC_final)

---
## Logistic CVEC/TVEC

log_cvec_pipe = Pipeline([
    ('cvec', CountVectorizer()),
    ('logr', LogisticRegression())
])

log_cvec_param = {
    'cvec__stop_words' : ['english', None],
    'cvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)],
    'logr__max_iter': range(50, 501, 50)
}


gs_log_cvec = GridSearchCV(log_cvec_pipe, param_grid = log_cvec_param)

gs_log_cvec.fit(XS_train, yS_train)

gs_log_cvec.best_params_

gs_log_cvec.score(XS_train, yS_train), gs_log_cvec.score(XS_test, yS_test), gs_log_cvec.score(XS_final, yS_final)

gs_log_cvec.score(XC_train, yC_train), gs_log_cvec.score(XC_test, yC_test), gs_log_cvec.score(XC_final, yC_final)

---


log_tvec_pipe = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('logr', LogisticRegression())
])

log_tvec_param = {
    'tvec__stop_words' : ['english', None],
    'tvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)],
    'logr__max_iter': range(50, 501, 50)
}


gs_log_tvec = GridSearchCV(log_tvec_pipe, param_grid = log_tvec_param)

gs_log_tvec.fit(XS_train, yS_train)

gs_log_tvec.best_params_

gs_log_tvec.score(XS_train, yS_train), gs_log_tvec.score(XS_test, yS_test), gs_log_tvec.score(XS_final, yS_final)

gs_log_tvec.score(XC_train, yC_train), gs_log_tvec.score(XC_test, yC_test), gs_log_tvec.score(XC_final, yC_final)

--- 
## ExtraTrees CVEC/TVEC

ext_cvec_pipe = Pipeline([
    ('cvec', CountVectorizer()),
    ('ext', ExtraTreesClassifier())
])

ext_cvec_param = {
    'cvec__stop_words' : ['english', None],
    'cvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)]
}

gs_ext_cvec = GridSearchCV(ext_cvec_pipe, param_grid = ext_cvec_param)

gs_ext_cvec.fit(XS_train, yS_train)

gs_ext_cvec.best_params_

gs_ext_cvec.score(XS_train, yS_train), gs_ext_cvec.score(XS_test, yS_test), gs_ext_cvec.score(XS_final, yS_final)

ext_tvec_pipe = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('rfc', RandomForestClassifier())
])

ext_tvec_param = {
    'tvec__stop_words' : ['english', None],
    'tvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)]
}

gs_ext_tvec = GridSearchCV(ext_tvec_pipe, param_grid = ext_tvec_param)

gs_ext_tvec.fit(XS_train, yS_train)

gs_ext_tvec.best_params_

gs_ext_tvec.score(XS_train, yS_train), gs_ext_tvec.score(XS_test, yS_test), gs_ext_tvec.score(XS_final, yS_final)

---
## GBoost CVEC/TVEC

gb_cvec_pipe = Pipeline([
    ('cvec', CountVectorizer()),
    ('gb', GradientBoostingClassifier())
])

gb_cvec_param = {
    'cvec__stop_words' : ['english', None],
    'cvec__ngram_range': [(1,1), (2,2), (3,3), (1,2),(1,3)],
    'gb__learning_rate': [0.05, 0.1, 0.15, 0.2],
    'gb__n_estimators': [200, 300]
}

gs_gb_cvec = GridSearchCV(gb_cvec_pipe, param_grid = gb_cvec_param)

gs_gb_cvec.fit(XS_train, yS_train)

gs_gb_cvec.best_params_

gs_gb_cvec.score(XS_train, yS_train), gs_gb_cvec.score(XS_test, yS_test), gs_gb_cvec.score(XS_final, yS_final)

---

gb_tvec_pipe = Pipeline([
    ('tvec', CountVectorizer()),
    ('gb', GradientBoostingClassifier())
])

gb_tvec_param = {
    'tvec__tokenizer': [None, my_lemma],
    'tvec__ngram_range': [(1,2),(1,3)],
    'gb__learning_rate': [0.1, 0.2],
    'gb__n_estimators': [150, 200]
}

gs_gb_tvec = GridSearchCV(gb_tvec_pipe, param_grid = gb_tvec_param)

gs_gb_tvec.fit(XS_train, yS_train)

gs_gb_tvec.best_params_

gs_gb_tvec.score(XS_train, yS_train), gs_gb_tvec.score(XS_test, yS_test), gs_gb_tvec.score(XS_final, yS_final)
