==> code/06_analysis.ipynb <==
import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer

df = pd.read_csv('../data/df_analysis.csv')

df

df[df['is_news'] != df['preds']]

def extract_bi_trigrams(dataframe, column, n):
    '''
    Extract top n bigrams and trigrams from dataframe[column]
    '''
    for i in range(1, 4):
        cvec = CountVectorizer(ngram_range = (i, i), stop_words = 'english')
        cvec.fit(dataframe[column])
        vectorized_df = pd.DataFrame(cvec.transform(dataframe[column]).todense(), columns = cvec.get_feature_names_out())
        print(vectorized_df.head().sum().sort_values(ascending = False).head(n))

extract_bi_trigrams(df, 'title', 30)

extract_bi_trigrams(df_sub, 'title', 30)

extract_bi_trigrams(df_com, 'body', 30)
